Do not catalog creation/modification date.
In Vifib those dates are not used.

There is issue with getting creation/modification date on big workflows, which
leads to (1) high memory consumption and (2) not freeing memory later.

Vifib workflows have ten thousands of entires (and they grow), so until proper
solution if found it is better to avoid often calculation of those dates.